Minor cleanup + move to Sparkle 1.21.2

+++ src/mac/up.mm
@@ -99,7 +99,7 @@ void Up::setAutomaticInterval(const int seconds)
 
 int Up::getAutomaticInterval()
 {
-    return [d->updater updateCheckInterval];
+    return static_cast<int>([d->updater updateCheckInterval]);
 }
